Finished Payload implementation
Nekrosis and user payload implemented. Need to change this to root access required section as it uses the cmd to open edge and call the payload

Going to work on iterating through the whole folder and then looking at the folder for startup app shortcuts as taskbars are stored seperately. Got a little off base with it.